Update versioning to use file modification date
Changed automatic versioning in both DesktopIconBackupManager.py and DesktopIconBackupManager.spec to derive the version from the last modification date of the source file, instead of the current date. This ensures the version only updates when the code is actually changed. Also updated the PyInstaller spec to use the new versioning scheme in the output filename.